#5 [FEAT] : execute naviagate action on fragments
