+++ week1/.idea/misc.xml
@@ -15,7 +15,6 @@
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/ic_miniuns.xml" value="0.3072916666666667" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/ic_profile.xml" value="0.221875" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/ic_setting.xml" value="0.221875" />
-        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/line_onboarding.xml" value="0.30625" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/login_btn_background.xml" value="0.22" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/rectangle_border_pink.xml" value="0.30625" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/drawable/rectangle_fill_gray.xml" value="0.30625" />
@@ -29,17 +28,16 @@
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/font/noto_sans_kr_regular.xml" value="1.2142857142857142" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_home.xml" value="0.22" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_main.xml" value="0.26992753623188404" />
-        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_on_boarding.xml" value="0.22" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_second.xml" value="0.26992753623188404" />
-        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_sign_in.xml" value="0.1" />
+        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_sign_in.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/activity_sign_up.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_empty_follower.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_empty_following.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_follower.xml" value="0.21920289855072464" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_home.xml" value="0.36666666666666664" />
-        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_on_boarding_fisrt.xml" value="0.14666666666666667" />
-        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_on_boarding_last.xml" value="0.2644927536231884" />
-        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_on_boarding_second.xml" value="0.2644927536231884" />
+        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_on_boarding_fisrt.xml" value="0.23822463768115942" />
+        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_on_boarding_last.xml" value="0.23822463768115942" />
+        <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_on_boarding_second.xml" value="0.23822463768115942" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_profile.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/fragment_repo.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/alsd2/AndroidStudioProjects/Android-Minwoo/week1/app/src/main/res/layout/item_follower_list.xml" value="0.4466666666666667" />
