+++ week1/app/src/main/java/com/example/week1/data/local/RepoData.kt
@@ -1,4 +1,4 @@
-package com.example.week1.data
+package com.example.week1.data.local
 
 data class RepoData(
     val repo : String,
