+++ week1/app/src/main/java/com/example/week1/data/remote/RequestLoginData.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.data.remote
 
 data class RequestLoginData(
     val email : String,
