+++ week1/app/src/main/java/com/example/week1/data/remote/RequestSignUpData.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.data.remote
 
 data class RequestSignUpData(
     val email : String,
