+++ week1/app/src/main/java/com/example/week1/data/remote/ResponseSignUpData.kt
@@ -1,14 +1,15 @@
-package com.example.week1
+package com.example.week1.data.remote
 
-data class ResponseLoginData(
+data class ResponseSignUpData(
     val status : Int,
-    val success : Boolean,
+    val success : String,
     val message : String,
     val data : Data
 ) {
     data class Data(
         val id : Int,
         val name : String,
+        val password : String,
         val email : String
     )
 }
