+++ week1/app/src/main/java/com/example/week1/data/remote/ServiceCreator.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.data.remote
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
