+++ week1/app/src/main/java/com/example/week1/ui/FollowerFragment.kt
@@ -1,18 +1,14 @@
 package com.example.week1.ui
 
-import android.content.Intent
 import android.os.Bundle
-import android.telecom.Call
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.week1.ui.adapter.FollowerAdapter
-import com.example.week1.data.FollowerData
+import com.example.week1.data.local.FollowerData
 import com.example.week1.databinding.FragmentFollowerBinding
