+++ week1/app/src/main/java/com/example/week1/ui/SignInActivity.kt
@@ -5,10 +5,13 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
+import com.example.week1.data.remote.RequestLoginData
+import com.example.week1.data.remote.ResponseLoginData
 import com.example.week1.databinding.ActivitySignInBinding
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import com.example.week1.data.remote.ServiceCreator
 
 class SignInActivity : AppCompatActivity() {
