+++ week1/app/src/main/res/layout/fragment_on_boarding_fisrt.xml
@@ -11,7 +11,7 @@
         android:id="@+id/iv_on_boarding"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="70dp"
+        android:layout_marginTop="40dp"
         android:src="@drawable/img_github"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -21,7 +21,7 @@
         android:id="@+id/tv_on_boarding_welcome"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="55dp"
+        android:layout_marginTop="40dp"
         android:text="SOPTHub에 오신걸 \n 환영합니다 "
         android:textAlignment="center"
         android:textColor="#F658A6"
@@ -39,7 +39,7 @@
 
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
-        android:layout_marginBottom="60dp"
+        android:layout_marginBottom="40dp"
         android:background="@drawable/login_btn_background"
         android:fontFamily="@font/noto_sans_kr"
         android:text="다음으로"
