+++ week1/app/src/main/res/layout/fragment_on_boarding_last.xml
@@ -21,7 +21,7 @@
         android:id="@+id/tv_on_boarding_start"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="100dp"
+        android:layout_marginTop="50dp"
         android:text="SOPTHub를\n 시작해볼까요 ?"
         android:textAlignment="center"
         android:textColor="#F658A6"
@@ -39,7 +39,7 @@
 
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
-        android:layout_marginBottom="60dp"
+        android:layout_marginBottom="40dp"
 
         android:background="@drawable/login_btn_background"
         android:fontFamily="@font/noto_sans_kr"
