+++ week1/app/src/main/res/layout/fragment_on_boarding_second.xml
@@ -11,7 +11,7 @@
         android:id="@+id/tv_on_boarding_goodtime"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="300dp"
+        android:layout_marginTop="200dp"
         android:text="SOPTHub에서 즐거운 시간\n 보내시길 바랍니다 !"
         android:textAlignment="center"
         android:textColor="#F658A6"
@@ -30,7 +30,7 @@
 
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
-        android:layout_marginBottom="60dp"
+        android:layout_marginBottom="40dp"
 
         android:background="@drawable/login_btn_background"
         android:fontFamily="@font/noto_sans_kr"
@@ -41,6 +41,6 @@
 
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
